Added Time based clearing of caches;

+++ P42.Utils/DownloadCache.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using System.Text;
 using System.IO;
 using System.Diagnostics;
+using System.Linq;
 
 namespace P42.Utils
 {
@@ -24,13 +25,20 @@ namespace P42.Utils
 
         static readonly object _locker = new object();
         static readonly Dictionary<string, Task<bool>> _downloadTasks = new Dictionary<string, Task<bool>>();
-        //static readonly System.Security.Cryptography.MD5 _md5 = System.Security.Cryptography.MD5.Create();
 
         public static string Download(string url, string folderName = null)
         {
             var task = Task.Run(() => DownloadAsync(url, folderName));
             return task.Result;
         }
+
+        public static List<string> List(string folderName)
+        {
+            var folderPath = FolderPath(folderName);
+            var files = System.IO.Directory.EnumerateFiles(folderPath);
+            return files.ToList();
+        }
+
 
 
         public static async Task<string> DownloadAsync(string url, string folderName = null)
@@ -53,6 +61,12 @@ namespace P42.Utils
         }
 
         public static bool Clear(string url = null, string folderName = null)
+            => Clear(DateTime.MinValue, url, folderName);
+
+        public static bool Clear(TimeSpan timeSpan, string url = null, string folderName = null)
+            => Clear(DateTime.Now - timeSpan, url, folderName);
+
+        public static bool Clear(DateTime dateTime, string url = null, string folderName = null)
         {
             if (string.IsNullOrWhiteSpace(url))
             {
@@ -61,18 +75,31 @@ namespace P42.Utils
                 if (System.IO.Directory.Exists(folderPath))
                 {
                     var files = System.IO.Directory.EnumerateFiles(folderPath);
+                    bool filesRemaining = false;
                     foreach (var file in files)
-                        System.IO.File.Delete(file);
-                    System.IO.Directory.Delete(folderPath);
-                    return true;
+                    {
+                        if (System.IO.File.Exists(file))
+                        {
+                            if (System.IO.File.GetLastWriteTime(file) < dateTime)
+                                System.IO.File.Delete(file);
+                            else
+                                filesRemaining = true;
+                        }
+                        if (!filesRemaining)
+                            System.IO.Directory.Delete(folderPath);
+                        return true;
+                    }
                 }
                 return false;
             }
             var path = CachedPath(url, folderName);
             if (!string.IsNullOrEmpty(path) && System.IO.File.Exists(path))
             {
-                System.IO.File.Delete(path);
-                return true;
+                if (System.IO.File.GetLastWriteTime(path) < dateTime)
+                {
+                    System.IO.File.Delete(path);
+                    return true;
+                }
             }
             return false;
         }
